+++ packages/terra-tabs/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased
+* Added
+  * Migrate docs to terra-framework-docs
 
 ## 6.60.0 - (November 16, 2021)